// src/lib/api/models/SkuResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SkuResponse = {
    id: number;
    order_product_code: string;
    invoice_product_code: string;
    description: string;
    package: string;
    commission_percentage: number;
    created: number;
    updated: (number | null);
};